+++ CHANGELOG.md
@@ -28,6 +28,8 @@ All changes to the **Hearts of Modding** extension will be documented in this fi
 
 - **String interner memory leak:** The interner stored every interned string forever. Implemented a garbage collector to solve this.
 
+- **Miscelaneous idea scanner fixes.**
+
 ## [v0.16.1] - 2026-06-04
 
 ### 🩹 Fixed